Overhaulted the alert controller again: Added various checks built-in to limit amount of user interaction required to process fully.
Still requires a significant amount of interaction due to a number of values near the end that appear to behave as though they aren't receiving time variables despite the value being present in the variable when printed throughout the code.  Will attempt to narrow down possibilities.

Added a few graphic elements to improve the aesthetics of the app environment.